Generalized side-conditions
This makes the Drinker's Problem formula (DP) not provable
and its Goedel-Gentzen translation provable.

Earlier the reverse was the case because the EVC was not
strong enough to express checks *inside* a single phase.
Thus, in the "simple" form of the DP,

   \E x. \A y. p(x) => p(y)

this was incorrectly triggering an EVC success.

Now every rule has an additional list of freshness tests
for the EVs that check that they are fresh in (the
substitutions of) term evars.