percolate: new algorithm that is fast and yet complete
For any selected fact S:

1. Run through all rules where S can match *any* premise, not
   just the left-most one. Call the specializations R.

2. Store all the R.

3. While R is non-empty:

   - Specialize R with the *left-most* premise matched from the
     active set.

   - Register any facts so produced

Note that since each selected fact is considered to be globally
fresh, the initial R set is also globally fresh. Hence, there is
no need to store any of the intermediate results in the while
loop in step 3: none of these rules will ever need to be
considered again.

Passes the current battery of tests.