Adding right-Delta things
There is also a flag in model.ml called "antecedent_first".
Setting it to true causes the right-Deltas to be basically
ignored (I think). Setting it to false goes into the
succedent of the implication-left branch first, which in
turn makes use of these right-Deltas to kill incompatible
assignments.

tests/classical/regression/REGR0019.mg tests all polarity
combinations that are relevant for this change.